reorder the creation of the base partitions / mdraid devs #131
If we create in the /boot swap / order then we more easily
allocate the remaining space on greater than minimum size
system disks. Ie many raid installs will be on greater than 8GB
system disks so we end up more reliably with all the space in
root rather than swap. Also space calculation for /boot and swap
are a great deal easier and tend not to change with system disk
size.