added links and label
Added _concepts label to the top of the page since it didn't have one.  The existing Rock-ons paragraph had a link in the last sentence, so used that as my model for consistency, and added a similar sentence to all the other sections, except Secure Shell, and Smart Manager.  could not find existing pages for those sections.  Changed all the links from hard coded html link to :ref: RST notation.