Add section to address importing a poorly pool #275
In some cases, such as disaster recovery, it is useful to be able to
import poorly pools, such as those that will only mount ro, into a fresh
install. This section provides the current work-around for enabling such
an import by detailing how to mount by-hand before an import so that
custom mount options can be enacted.