Add sections for missing services #460
The list of services in our docs does not completely match the
list present in Rockstor's UI: System > Services page. In the
docs, some are missing, while the rest is not listed in the same
order.

This commit thus ensure a match in the services listed and their
order between our docs and Rockstor UI.