smartmontools.org linkcheck failures #533
Temporarily use alternative URLs for 2 key manual page links.
Add domain to linkcheck_ignore config, with a proviso that
this be temporary.

Smartmontools the project, looks to be undergoing a
large migration from their long time host of
Sourceforge (SVN) to GitHub (git). It is assumed
that the website difficulties are related to this
move.